refactor: move schemas to a dedicated package
